+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ## [Unreleased]
 
  * Support PHP version 8.0. [#74]
+ * Revert fix for Gmagick bug `No encode delegate for this image format`. [#70]
 
 ## [1.0.2] (2020-06-13)